First working version

+++ dk51/AMahnyov/lab1/biclycleMath.h
@@ -1,8 +1,8 @@
 #ifndef BICLYCLEMATH_H_INCLUDED
 #define BICLYCLEMATH_H_INCLUDED
 
-int bicycleSum(int iterations);
-int bicycleModulo(int input);
-int bicyclePow(int num, int power);
+double bicycleSum(int iterations);
+double bicycleModulo(int input);
+double bicyclePow(int num, int power);
 
 #endif // BICLYCLEMATH_H_INCLUDED
